+++ app/src/main/java/com/coupOn/platform/coupOn/Model/Coupon.java
@@ -14,13 +14,13 @@ public class Coupon
     private String interest;
     private String discountType;
     private String code;
-    private double price;
+    private int price;
     private int rank;
 
     private Uri uri;
 
     public Coupon(String couponImage, String couponName, String expireDate, String location, String description
-            , String ownerId, String couponId, String interest, String discountType, String code, int rank, double price) {
+            , String ownerId, String couponId, String interest, String discountType, String code, int rank, int price) {
         this.couponImage = couponImage;
         this.couponName = couponName;
         this.expireDate = expireDate;
@@ -54,7 +54,7 @@ public class Coupon
                 '}';
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
